user/requirements.txt
Flask==2.0.3
psycopg2==2.9.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
# ariadne==0.14.1
pytest==7.1.1
strawberry-graphql==0.123.3
